KIEV21DEVS-1231 Fixed most of the errors

+++ backend/src/main/java/com/eshop/backend/product/dao/models/ProductModel.java
@@ -14,7 +14,7 @@ public class ProductModel {
 
     private String productName;
 
-    private Long productAmount;
+    private int productAmount;
 
     private double productPrice;
